feat: text transformations for text-based components
